[refactor/#224] myActiviesApi.ts 파일명 오타 수정 및 import 경로 수정
